fix: move Waku related methods to waku_node module
