+++ src/libs/common.py
@@ -38,18 +38,3 @@ def delay(num_seconds):
 
 def gen_step_id():
     return f"{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}__{str(uuid.uuid4())}"
-
-
-def peer_info2id(peer, is_nwaku=True):
-    return peer_info2multiaddr(peer, is_nwaku).split("/")[-1]
-
-
-def peer_info2multiaddr(peer, is_nwaku=True):
-    if is_nwaku:
-        return peer["multiaddr"]
-    else:
-        return peer["multiaddrs"][0]
-
-
-def multiaddr2id(multiaddr):
-    return multiaddr.split("/")[-1]
